requirements.txt
rich>=13.7.0
pyyaml>=6.0.1
python-ffmpeg>=2.0.10